#18 refactor : MyPage 패키지 구조 수정
