--- Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/EditProfileUiEvent.kt
@@ -1,5 +1,0 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage
-
-sealed interface EditProfileUiEvent {
-
-}
