+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/edit_profile/EditProfileViewModel.kt
@@ -1,4 +1,4 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage
+package com.avengers.nibobnebob.presentation.ui.main.mypage.edit_profile
 
 import androidx.lifecycle.ViewModel
