+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/my_list/MyRestaurantListFragment.kt
@@ -1,4 +1,4 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage
+package com.avengers.nibobnebob.presentation.ui.main.mypage.my_list
 
 import android.os.Bundle
 import android.view.View
@@ -6,11 +6,14 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
 import com.avengers.nibobnebob.R
-import com.avengers.nibobnebob.databinding.FragmentEditProfileBinding
+import com.avengers.nibobnebob.databinding.FragmentMyRestaurantListBinding
 import com.avengers.nibobnebob.presentation.base.BaseFragment
+import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedViewModel
+import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedUiEvent
 
-class EditProfileFragment :
-    BaseFragment<FragmentEditProfileBinding>(R.layout.fragment_edit_profile) {
+class MyRestaurantListFragment :
+    BaseFragment<FragmentMyRestaurantListBinding>(R.layout.fragment_my_restaurant_list) {
+
     private val sharedViewModel: MyPageSharedViewModel by viewModels()
     private lateinit var navController: NavController
 
@@ -29,12 +32,10 @@ class EditProfileFragment :
         viewLifecycleOwner.repeatOnStarted {
             sharedViewModel.uiEvent.collect { event ->
                 when (event) {
-                    is MyPageUiEvent.NavigateToBack ->
+                    is MyPageSharedUiEvent.NavigateToBack ->
                         navController.navigateUp()
 
-                    else -> {
-                        Unit
-                    }
+                    else -> Unit
                 }
 
             }
