+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/share/MyPageSharedUiEvent.kt
@@ -1,0 +1,9 @@
+package com.avengers.nibobnebob.presentation.ui.main.mypage.share
+
+sealed interface MyPageSharedUiEvent {
+    data object NavigateToEditProfile : MyPageSharedUiEvent
+    data object NavigateToMyList : MyPageSharedUiEvent
+    data object NavigateToWishList : MyPageSharedUiEvent
+    data object NavigateToBack : MyPageSharedUiEvent
+
+}
