+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/main/mypage/wish_list/WishRestaurantListFragment.kt
@@ -1,4 +1,4 @@
-package com.avengers.nibobnebob.presentation.ui.main.mypage
+package com.avengers.nibobnebob.presentation.ui.main.mypage.wish_list
 
 import android.os.Bundle
 import android.view.View
@@ -6,11 +6,13 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
 import com.avengers.nibobnebob.R
-import com.avengers.nibobnebob.databinding.FragmentMyRestaurantListBinding
+import com.avengers.nibobnebob.databinding.FragmentWishRestaurantListBinding
 import com.avengers.nibobnebob.presentation.base.BaseFragment
+import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedViewModel
+import com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedUiEvent
 
-class MyRestaurantListFragment :
-    BaseFragment<FragmentMyRestaurantListBinding>(R.layout.fragment_my_restaurant_list) {
+class WishRestaurantListFragment :
+    BaseFragment<FragmentWishRestaurantListBinding>(R.layout.fragment_wish_restaurant_list) {
 
     private val sharedViewModel: MyPageSharedViewModel by viewModels()
     private lateinit var navController: NavController
@@ -30,7 +32,7 @@ class MyRestaurantListFragment :
         viewLifecycleOwner.repeatOnStarted {
             sharedViewModel.uiEvent.collect { event ->
                 when (event) {
-                    is MyPageUiEvent.NavigateToBack ->
+                    is MyPageSharedUiEvent.NavigateToBack ->
                         navController.navigateUp()
 
                     else -> Unit
