+++ Aos/app/src/main/res/layout/fragment_my_page.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="vm"
-            type="com.avengers.nibobnebob.presentation.ui.main.mypage.MyPageSharedViewModel" />
+            type="com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedViewModel" />
         <variable
             name="ui_state"
             type="com.avengers.nibobnebob.presentation.ui.main.mypage.MyPageUiState" />
