+++ Aos/app/src/main/res/layout/fragment_wish_restaurant_list.xml
@@ -6,13 +6,13 @@
     <data>
         <variable
             name="vm"
-            type="com.avengers.nibobnebob.presentation.ui.main.mypage.MyPageSharedViewModel" />
+            type="com.avengers.nibobnebob.presentation.ui.main.mypage.share.MyPageSharedViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.main.mypage.WishRestaurantListFragment">
+        tools:context=".presentation.ui.main.mypage.wish_list.WishRestaurantListFragment">
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/gl_start"
