+++ Aos/app/src/main/res/navigation/nav_graph.xml
@@ -37,14 +37,14 @@
         app:popUpToInclusive="false"/>
     <fragment
         android:id="@+id/editProfileFragment"
-        android:name="com.avengers.nibobnebob.presentation.ui.main.mypage.EditProfileFragment"
+        android:name="com.avengers.nibobnebob.presentation.ui.main.mypage.edit_profile.EditProfileFragment"
         android:label="EditProfileFragment" />
     <fragment
         android:id="@+id/myRestaurantListFragment"
-        android:name="com.avengers.nibobnebob.presentation.ui.main.mypage.MyRestaurantListFragment"
+        android:name="com.avengers.nibobnebob.presentation.ui.main.mypage.my_list.MyRestaurantListFragment"
         android:label="MyRestaurantListFragment" />
     <fragment
         android:id="@+id/wishRestaurantListFragment"
-        android:name="com.avengers.nibobnebob.presentation.ui.main.mypage.WishRestaurantListFragment"
+        android:name="com.avengers.nibobnebob.presentation.ui.main.mypage.wish_list.WishRestaurantListFragment"
         android:label="WishRestaurantListFragment" />
 </navigation>
